Don't use `lastCursorMotionWasMouse` for detection.
For some obscure reason, this value gets reset to _true_ when exiting out of the shop menu and then reopening it, even if the entire sequence is done with a controller. It's not reliable.

Instead, switch to a method that checks whether the held buttons are controller buttons, regardless of what the "last cursor motion" was.