Use alternate "search parameters" in transpiler to avoid relying on the call to `tryToPurchaseItem`.
Since another mod is overwriting that entire instruction, we can use one that is less likely to be overwritten, which clamps the quantity to the amount available.

This does not change the actual transpiler output, only the method it uses to obtain semantic information about the underlying implementation.

Also bumps version to 1.0.0, since this has been stable for a long time with no bug reports, and updates minimum game/API versions to those that are currently tested.

Fixes #1